requirements.txt
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.22.1
cffi==1.16.0
click==8.1.7
confluent-kafka==2.3.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.4
pycares==4.4.0
pycparser==2.21
python-engineio==4.8.2
python-socketio==5.11.0
simple-websocket==1.0.0
typing_extensions==4.9.0
uvicorn==0.26.0
wsproto==1.2.0
yarl==1.9.4
